backend/app/config/mongo_db.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv


# Load environment variables from a .env file
load_dotenv()

# Get the MongoDB URI from the environment variables (set in Vercel)
MONGODB_URI_FASTAPI = os.getenv("MONGODB_URI_FASTAPI")

# Create a new MongoDB client and connect to the server using the URI
client = MongoClient(MONGODB_URI_FASTAPI)

db = client["ai_powered_interview_question_db"]
users_collection = db["users"]
