backend/requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
openai==1.63.2
python-dotenv==1.0.1
pymongo==4.11.1
pydantic[email]==2.10.6
fitz==0.0.1.dev2
tools==0.1.9
python-docx==1.1.2
# pymupdf==1.23.7